requirements.txt
eth-ape
ape-solidity
web3
pytest
python-dotenv